backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
motor
python-jose[cryptography]
passlib>=1.7.4
bcrypt>=4.3.0
email-validator
python-multipart
pytest
pytest-asyncio
httpx==0.27.2
anyio
testcontainers